+++ package.json
@@ -40,6 +40,7 @@
         "@heroicons/react": "^2.2.0",
         "firebase": "^12.7.0",
         "framer-motion": "^11.18.2",
+        "prop-types": "^15.8.1",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
         "react-helmet-async": "^2.0.5",
