+++ src/docs/PROJECT_TOTAL_CONSOLIDATED.md
@@ -1,6 +1,6 @@
 # 📚 Informe Consolidado Total - Proyecto MyAppGlass
 
-**Fecha de Consolidación:** 26 de Diciembre de 2025  
+**Fecha de Consolidación:** 27 de Diciembre de 2025  
 **Estado del Proyecto:** ✅ Estable, Refactorizado y Optimizado  
 **Arquitectura:** Clean Architecture / Feature-Based  
 **Tecnologías:** React 18, Vite, Chakra UI, Firebase
@@ -86,6 +86,12 @@ La lógica compleja se extrae de los componentes:
 ### 3. Consolidación de Documentación
 
 -   Creación de este informe maestro para alimentar herramientas de análisis de IA.
+
+### 4. Refactorización de Componentes Visuales (Clean Code)
+
+-   **Modularización**: Se dividieron componentes grandes como `ProjectDetailModal` y `Gallery` en subcomponentes especializados (`MapViewer`, `ProjectInfo`, `VisualViewer`, `GalleryThumbnails`), mejorando la legibilidad y el mantenimiento.
+-   **Robustez**: Implementación de `prop-types` para validación de datos en tiempo de ejecución.
+-   **Estandarización**: Renombrado de hooks a `camelCase` (`useIsMobile.js`, `useReclamoForm.js`) para cumplir con los estándares de la industria.
 
 ---
